@@ -1,53 +1,75 @@
 {
     "name": "doctrine/doctrine-mongo-odm-module",
-    "description": "Zend Framework Module that provides Doctrine MongoDB ODM functionality",
-    "type": "library",
+    "description": "Laminas Module which provides Doctrine MongoDB ODM functionality",
     "keywords": [
         "doctrine",
-        "odm",
+        "laminas",
         "mongodb",
         "module",
-        "zf"
+        "odm"
     ],
-    "homepage": "http://www.doctrine-project.org/",
+    "homepage": "https://www.doctrine-project.org/",
     "license": "MIT",
     "authors": [
-        { "name": "Kyle Spraggs", "email": "theman@spiffyjr.me" },
-        { "name": "Marco Pivetta", "email": "ocramius@gmail.com" },
-        { "name": "Evan Coury", "email": "me@evancoury.com" },
-        { "name": "Guilherme Blanco", "email": "guilhermeblanco@hotmail.com" },
-        { "name": "Tim Roediger", "email": "superdweebie@gmail.com" },
-        { "name": "Maciej Malarz", "email": "malarzm@gmail.com" },
-        { "name": "Andreas Braun", "email": "alcaeus@alcaeus.org" }
+        {
+            "name": "Kyle Spraggs",
+            "email": "theman@spiffyjr.me"
+        },
+        {
+            "name": "Marco Pivetta",
+            "email": "ocramius@gmail.com"
+        },
+        {
+            "name": "Evan Coury",
+            "email": "me@evancoury.com"
+        },
+        {
+            "name": "Guilherme Blanco",
+            "email": "guilhermeblanco@hotmail.com"
+        },
+        {
+            "name": "Tim Roediger",
+            "email": "superdweebie@gmail.com"
+        },
+        {
+            "name": "Maciej Malarz",
+            "email": "malarzm@gmail.com"
+        },
+        {
+            "name": "Andreas Braun",
+            "email": "alcaeus@alcaeus.org"
+        },
+        {
+            "name": "Tom H Anderson",
+            "email": "tom.h.anderson@gmail.com"
+        }
     ],
     "extra": {
-        "branch-alias": {
-            "dev-master": "1.1.x-dev"
-        },
-        "zf": {
+        "laminas": {
             "module": "DoctrineMongoODMModule"
         }
     },
     "require": {
-        "php": "^7.1",
-        "doctrine/doctrine-module": "^1.2 || ^2.1.7",
+        "php": "^7.2",
+        "alcaeus/mongo-php-adapter": "^1.1",
+        "doctrine/doctrine-module": "^4.0",
         "doctrine/mongodb-odm": "^1.1",
-        "zendframework/zend-mvc": "^2.7.15 || ^3.0.1",
-        "zendframework/zend-servicemanager": "^2.7.6 || ^3.1",
-        "zendframework/zend-stdlib": "^3.2.1",
-        "zendframework/zend-hydrator": "^2.2"
+        "laminas/laminas-hydrator": "^3.0.2",
+        "laminas/laminas-mvc": "^3.1.1",
+        "laminas/laminas-servicemanager": "^3.4",
+        "laminas/laminas-stdlib": "^3.2.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5",
-        "squizlabs/php_codesniffer": "^3.0.0",
-        "zendframework/zend-console": "^2.6",
-        "zendframework/zend-i18n": "^2.7.3",
-        "zendframework/zend-log": "^2.9",
-        "zendframework/zend-modulemanager": "^2.7.2",
-        "zendframework/zend-serializer": "^2.8",
-        "zendframework/zend-session": "^2.8.5",
-        "zendframework/zend-view": "^2.8.1",
-        "zendframework/zend-mvc-console": "^1.1.8"
+        "doctrine/coding-standard": "^7.0.2",
+        "laminas/laminas-console": "^2.8",
+        "laminas/laminas-i18n": "^2.10.1",
+        "laminas/laminas-log": "^2.12",
+        "laminas/laminas-modulemanager": "^2.8.4",
+        "laminas/laminas-mvc-console": "^1.2",
+        "laminas/laminas-serializer": "^2.9.1",
+        "laminas/laminas-session": "^2.9.1",
+        "laminas/laminas-view": "^2.11.4",
+        "phpunit/phpunit": "^8.5.2"
     },
     "autoload": {
         "psr-4": {
