@@ -1,18 +1,19 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModule\Collector;
 
-use ZendDeveloperTools\Collector\CollectorInterface;
-use ZendDeveloperTools\Collector\AutoHideInterface;
-
-use Zend\Mvc\MvcEvent;
-
 use DoctrineMongoODMModule\Logging\DebugStack;
+use Laminas\DeveloperTools\Collector\AutoHideInterface;
+use Laminas\DeveloperTools\Collector\CollectorInterface;
+use Laminas\Mvc\MvcEvent;
+use function count;
 
 /**
- * Collector to be used in ZendDeveloperTools to record and display Mongo
+ * Collector to be used in Laminas\DeveloperTools to record and display Mongo
  * queries
  *
- * @license MIT
  * @link    www.doctrine-project.org
  */
 class MongoLoggerCollector implements CollectorInterface, AutoHideInterface
@@ -20,26 +21,18 @@ class MongoLoggerCollector implements CollectorInterface, AutoHideInterface
     /**
      * Collector priority
      */
-    const PRIORITY = 10;
+    public const PRIORITY = 10;
 
-    /**
-     * @var DebugStack
-     */
+    /** @var DebugStack */
     protected $mongoLogger;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $name;
 
-    /**
-     * @param DebugStack $mongoLogger
-     * @param string     $name
-     */
-    public function __construct(DebugStack $mongoLogger, $name)
+    public function __construct(DebugStack $mongoLogger, string $name)
     {
         $this->mongoLogger = $mongoLogger;
-        $this->name = (string) $name;
+        $this->name        = (string) $name;
     }
 
     /**
@@ -55,7 +48,7 @@ class MongoLoggerCollector implements CollectorInterface, AutoHideInterface
      */
     public function getPriority()
     {
-        return static::PRIORITY;
+        return self::PRIORITY;
     }
 
     /**
@@ -73,26 +66,20 @@ class MongoLoggerCollector implements CollectorInterface, AutoHideInterface
         return empty($this->mongoLogger->queries);
     }
 
-    /**
-     * @return int
-     */
-    public function getQueryCount()
+    public function getQueryCount() : int
     {
         return count($this->mongoLogger->queries);
     }
 
     /**
-     * @return array
+     * @return mixed[]
      */
-    public function getQueries()
+    public function getQueries() : array
     {
         return $this->mongoLogger->queries;
     }
 
-    /**
-     * @return float
-     */
-    public function getQueryTime()
+    public function getQueryTime() : float
     {
         return 0.0;
     }
