@@ -1,15 +1,15 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModule\Options;
 
-use Zend\Stdlib\AbstractOptions;
+use Laminas\Stdlib\AbstractOptions;
 
 /**
  * Document manager options for doctrine mongo
  *
- * @license MIT
  * @link    http://www.doctrine-project.org/
- * @since   0.1.0
- * @author  Tim Roediger <superdweebie@gmail.com>
  */
 class DocumentManager extends AbstractOptions
 {
@@ -41,60 +41,47 @@ class DocumentManager extends AbstractOptions
     protected $eventManager = 'odm_default';
 
     /**
-     *
-     * @param type $configuration
-     * @return \DoctrineMongoODMModule\Options\DocumentManager
+     * @param mixed $configuration
      */
-    public function setConfiguration($configuration)
+    public function setConfiguration($configuration) : self
     {
         $this->configuration = (string) $configuration;
+
         return $this;
     }
 
-    /**
-     * @return string
-     */
-    public function getConfiguration()
+    public function getConfiguration() : string
     {
-        return "doctrine.configuration.{$this->configuration}";
+        return 'doctrine.configuration.' . $this->configuration;
     }
 
     /**
-     *
-     * @param type $connection
-     * @return \DoctrineMongoODMModule\Options\DocumentManager
+     * @param mixed $connection
      */
-    public function setConnection($connection)
+    public function setConnection($connection) : self
     {
         $this->connection = (string) $connection;
+
         return $this;
     }
 
-    /**
-     * @return string
-     */
-    public function getConnection()
+    public function getConnection() : string
     {
-        return "doctrine.connection.{$this->connection}";
+        return 'doctrine.connection.' . $this->connection;
+    }
+
+    public function getEventManager() : string
+    {
+        return 'doctrine.eventmanager.' . $this->eventManager;
     }
 
     /**
-     *
-     * @return string
+     * @param mixed $eventManager
      */
-    public function getEventManager()
-    {
-        return "doctrine.eventmanager.{$this->eventManager}";
-    }
-
-    /**
-     *
-     * @param type $eventManager
-     * @return \DoctrineMongoODMModule\Options\DocumentManager
-     */
-    public function setEventManager($eventManager)
+    public function setEventManager($eventManager) : self
     {
         $this->eventManager = (string) $eventManager;
+
         return $this;
     }
 }
