@@ -1,27 +1,40 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModule\Service;
 
-use Interop\Container\ContainerInterface;
 use DoctrineModule\Service\AbstractFactory as DoctrineModuleAbstractFactory;
+use Interop\Container\ContainerInterface;
+use Laminas\Stdlib\AbstractOptions;
 use RuntimeException;
+use function sprintf;
 
+// phpcs:disable SlevomatCodingStandard.Classes.SuperfluousAbstractClassNaming
 abstract class AbstractFactory extends DoctrineModuleAbstractFactory
 {
-    public function getOptions(ContainerInterface $container, $key, $name = null)
+// phpcs:enable SlevomatCodingStandard.Classes.SuperfluousAbstractClassNaming
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getOptions(ContainerInterface $container, $key, $name = null) : AbstractOptions
     {
         if ($name === null) {
             $name = $this->getName();
         }
 
-        $options = $container->get('config');
-        $options = $options['doctrine'];
-        if ($mappingType = $this->getMappingType()) {
+        $options     = $container->get('config');
+        $options     = $options['doctrine'];
+        $mappingType = $this->getMappingType();
+
+        if ($mappingType) {
             $options = $options[$mappingType];
         }
-        $options = isset($options[$key][$name]) ? $options[$key][$name] : null;
 
-        if (null === $options) {
+        $options = $options[$key][$name] ?? null;
+
+        if ($options === null) {
             throw new RuntimeException(
                 sprintf(
                     'Options with name "%s" could not be found in "doctrine.%s".',
