@@ -1,18 +1,18 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModule\Service;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
 use DoctrineMongoODMModule\Options;
 use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 /**
  * Factory creates a mongo document manager
  *
- * @license MIT
  * @link    http://www.doctrine-project.org/
- * @since   0.1.0
- * @author  Tim Roediger <superdweebie@gmail.com>
  */
 class DocumentManagerFactory extends AbstractFactory
 {
@@ -21,15 +21,19 @@ class DocumentManagerFactory extends AbstractFactory
      *
      * @return DocumentManager
      */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, ?array $options = null)
     {
         $options      = $this->getOptions($container, 'documentmanager');
         $connection   = $container->get($options->getConnection());
         $config       = $container->get($options->getConfiguration());
         $eventManager = $container->get($options->getEventManager());
+
         return DocumentManager::create($connection, $config, $eventManager);
     }
 
+    /**
+     * @return mixed
+     */
     public function createService(ServiceLocatorInterface $container)
     {
         return $this($container, DocumentManager::class);
@@ -37,10 +41,8 @@ class DocumentManagerFactory extends AbstractFactory
 
     /**
      * Get the class name of the options associated with this factory.
-     *
-     * @return string
      */
-    public function getOptionsClass()
+    public function getOptionsClass() : string
     {
         return Options\DocumentManager::class;
     }
