@@ -1,24 +1,22 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModuleTest\Assets;
 
+use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\Repository\RepositoryFactory;
 
-/**
- * Class CustomRepositoryFactory
- * @package DoctrineMongoODMModuleTest\Assets
- */
 class CustomRepositoryFactory implements RepositoryFactory
 {
     /**
      * Stub.
      *
      * @param DocumentManager $documentManager The DocumentManager instance.
-     * @param string $documentName The name of the document.
-     *
-     * @return \Doctrine\Common\Persistence\ObjectRepository
+     * @param string          $documentName    The name of the document.
      */
-    public function getRepository(DocumentManager $documentManager, $documentName)
+    public function getRepository(DocumentManager $documentManager, $documentName) : ObjectRepository
     {
     }
 }
