@@ -1,12 +1,11 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModuleTest\Assets;
 
 use Doctrine\ODM\MongoDB\Types\IntType;
 
-/**
- * Class CustomType
- * @package DoctrineMongoODMModuleTest\Assets
- */
 class CustomType extends IntType
 {
 }
