@@ -1,13 +1,16 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModuleTest\Doctrine;
 
 use DoctrineMongoODMModuleTest\AbstractTest;
 
 class AnnotationTest extends AbstractTest
 {
-    public function testAnnotation()
+    public function testAnnotation() : void
     {
         $documentManager = $this->getDocumentManager();
-        $metadata = $documentManager->getClassMetadata('DoctrineMongoODMModuleTest\Assets\Document\Annotation');
+        $metadata        = $documentManager->getClassMetadata('DoctrineMongoODMModuleTest\Assets\Document\Annotation');
     }
 }
