@@ -1,14 +1,16 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModuleTest\Doctrine;
 
 use DoctrineMongoODMModuleTest\AbstractTest;
 
 class AuthenticationAdapterFactoryTest extends AbstractTest
 {
-    public function testAuthenticationAdapterFactory()
+    public function testAuthenticationAdapterFactory() : void
     {
-
         $adapter = $this->serviceManager->get('doctrine.authenticationadapter.odm_default');
-        $this->assertInstanceOf('Zend\Authentication\Adapter\AdapterInterface', $adapter);
+        $this->assertInstanceOf('Laminas\Authentication\Adapter\AdapterInterface', $adapter);
     }
 }
