@@ -1,19 +1,22 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModuleTest\Doctrine;
 
+use Doctrine\Laminas\Hydrator\DoctrineObject;
 use Doctrine\ODM\MongoDB\DocumentManager;
-use DoctrineModule\Stdlib\Hydrator\DoctrineObject;
 use DoctrineMongoODMModule\Service\DoctrineObjectHydratorFactory;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 use PHPUnit\Framework\TestCase;
-use Zend\Hydrator\HydratorPluginManager;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use PHPUnit_Framework_MockObject_MockObject;
 
 class DoctrineObjectHydratorFactoryTest extends TestCase
 {
-    /** @var ServiceLocatorInterface|\PHPUnit_Framework_MockObject_MockObject */
+    /** @var ServiceLocatorInterface|PHPUnit_Framework_MockObject_MockObject */
     protected $services;
 
-    protected function setUp()
+    protected function setUp() : void
     {
         parent::setUp();
 
@@ -26,28 +29,10 @@ class DoctrineObjectHydratorFactoryTest extends TestCase
             ->willReturn($this->prophesize(DocumentManager::class)->reveal());
     }
 
-    public function testReturnsHydratorInstance()
+    public function testReturnsHydratorInstance() : void
     {
         $factory  = new DoctrineObjectHydratorFactory();
         $hydrator = $factory($this->services, DoctrineObject::class);
-
-        $this->assertInstanceOf(DoctrineObject::class, $hydrator);
-    }
-
-    public function testReturnsHydratorInstanceV2()
-    {
-        /** @var HydratorPluginManager $hydratorPluginManager */
-        $hydratorPluginManager = $this->getMockBuilder(HydratorPluginManager::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $hydratorPluginManager
-            ->expects($this->once())
-            ->method('getServiceLocator')
-            ->willReturn($this->services);
-
-        $factory  = new DoctrineObjectHydratorFactory();
-        $hydrator = $factory->createService($hydratorPluginManager);
 
         $this->assertInstanceOf(DoctrineObject::class, $hydrator);
     }
