@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModuleTest\Doctrine;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
@@ -6,7 +9,7 @@ use DoctrineMongoODMModuleTest\AbstractTest;
 
 final class DocumentManagerTest extends AbstractTest
 {
-    public function testDocumentManager()
+    public function testDocumentManager() : void
     {
         $documentManager = $this->getDocumentManager();
 
