@@ -1,27 +1,30 @@
 <?php
+
+declare(strict_types=1);
+
 namespace DoctrineMongoODMModuleTest\Doctrine;
 
 use DoctrineMongoODMModule\Module;
+use Laminas\EventManager\Event;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Application;
-use Zend\EventManager\Event;
 
 class ModuleTest extends TestCase
 {
-    public function testOdmDefaultIsUsedAsTheDocumentManagerIfNoneIsProvided()
+    public function testOdmDefaultIsUsedAsTheDocumentManagerIfNoneIsProvided() : void
     {
         $documentManager = $this->getMockbuilder('Doctrine\ODM\MongoDB\DocumentManager')
             ->disableOriginalConstructor()
             ->getMock();
 
-        $serviceManager = $this->createMock('Zend\ServiceManager\ServiceManager');
+        $serviceManager = $this->createMock('Laminas\ServiceManager\ServiceManager');
         $serviceManager->expects($this->once())
             ->method('get')
             ->with('doctrine.documentmanager.odm_default')
             ->will($this->returnValue($documentManager));
 
         $application = new Application();
-        $event = new Event('loadCli.post', $application, ['ServiceManager' => $serviceManager]);
+        $event       = new Event('loadCli.post', $application, ['ServiceManager' => $serviceManager]);
 
         $module = new Module();
         $module->loadCli($event);
@@ -29,7 +32,7 @@ class ModuleTest extends TestCase
         $this->assertSame($documentManager, $application->getHelperSet()->get('documentManager')->getDocumentManager());
     }
 
-    public function testDocumentManagerUsedCanBeSpecifiedInCommandLineArgument()
+    public function testDocumentManagerUsedCanBeSpecifiedInCommandLineArgument() : void
     {
         $argvBackup = $_SERVER['argv'];
 
@@ -37,14 +40,14 @@ class ModuleTest extends TestCase
             ->disableOriginalConstructor()
             ->getMock();
 
-        $serviceManager = $this->createMock('Zend\ServiceManager\ServiceManager');
+        $serviceManager = $this->createMock('Laminas\ServiceManager\ServiceManager');
         $serviceManager->expects($this->once())
             ->method('get')
             ->with('doctrine.documentmanager.some_other_name')
             ->will($this->returnValue($documentManager));
 
         $application = new Application();
-        $event = new Event('loadCli.post', $application, ['ServiceManager' => $serviceManager]);
+        $event       = new Event('loadCli.post', $application, ['ServiceManager' => $serviceManager]);
 
         $_SERVER['argv'][] = '--documentmanager=some_other_name';
 
